refactor dpm wave lookup routes
